[netsim] Calculate the delay for each packet.
For the + 900ms delay packets got delayed further and further. This
is now fixed by comparing the current time to the packet time and then
deciding to let the packet through immediately (if it was delayed for
too long) or calculating how much more time it should stay in the
queue.