[benchmark] Use POSIX setitimer for timing...
We are not interested in timing the scheduling of the system but
only how much time we spend in executing. This is the only metric
that will work for everyone.

This means that we can not measure the latency (e.g. waiting on
opening a TCP connection) but we can measure the latency of the
setup but we get a stable number regardles of the scheduling of
the system.

We might want to reintroduce the "wall clock" time again to have
two measures available.