Add a scrolling test for the painting time only.
Add a new test for the painting time only. This gives an estimation
of how much overhead there is outside painting by comparing with the
other test.

On the N900, this test has a lower variance compared to the other.
On desktop, it is a bit too fast for some page, the results are less
interesting in that case.