Add a class to write the output of the benchmark. Add output as CSV.
In order to output in different format and to different target, the
class BenchmarkOutputWriter was added. An instance of this class
is used to write the output of the benchmark.

Two subclasses were written: BenchmarkOutputHuman to output as the
old function, human readable results.
And BenchmarkOutputCsv, which output the results as CSV.

Reviewed-by: Viatcheslav Ostapenko