Update the robustness test to avoid visiting the same repeatedly website
The robustness test had a tendency to come back to the start page. When
this create a huge load, our IP is banned from the server, or CDN.

This patch introduce a list of visited addresses, and a list of
addresses to visit. This way, the test do not "attack" any website.